Moves unit test code to correct renderer. 🤦
